refactor(community): 패키지 구조 변경
- command/query 패키지 삭제
