+++ src/main/java/com/somemore/community/service/CreateCommunityBoardService.java
@@ -1,9 +1,9 @@
-package com.somemore.community.service.command;
+package com.somemore.community.service;
 
 import com.somemore.community.domain.CommunityBoard;
 import com.somemore.community.dto.request.CommunityBoardCreateRequestDto;
 import com.somemore.community.repository.CommunityBoardRepository;
-import com.somemore.community.usecase.command.CreateCommunityBoardUseCase;
+import com.somemore.community.usecase.CreateCommunityBoardUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
