+++ src/main/java/com/somemore/community/usecase/UpdateCommunityBoardUseCase.java
@@ -1,4 +1,4 @@
-package com.somemore.community.usecase.command;
+package com.somemore.community.usecase;
 
 import com.somemore.community.dto.request.CommunityBoardUpdateRequestDto;
